api/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
git-filter-repo==2.38.0
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.7
langchain-community==0.2.7
langchain-core==0.2.13
langchain-openai==0.1.15
langchain-text-splitters==0.2.2
langsmith==0.1.85
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.24.4
oauthlib==3.2.2
openai==1.35.13
openpyxl==3.1.5
orjson==3.10.6
outcome==1.3.0.post0
packaging==24.1
pandas==2.0.3
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
PyMySQL==1.1.1
pyparsing==3.1.2
pyproj==3.5.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
selenium==4.22.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.31
starlette==0.37.2
tenacity==8.5.0
tiktoken==0.7.0
timedelta==2020.12.3
tqdm==4.66.4
trio==0.26.0
trio-websocket==0.11.1
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
webdriver-manager==4.0.1
websocket-client==1.8.0
websockets==12.0
wsproto==1.2.0
yarl==1.9.4
